requirements.txt
작성예정

torch
torchvision
ffmpeg-python
opencv-python
Pillow